+++ lib/presentation/bindings/main_layout_binding.dart
@@ -5,6 +5,7 @@ import '../controllers/order_controller.dart';
 import '../controllers/wave_controller.dart';
 import '../controllers/customer_controller.dart';
 import '../controllers/product_controller.dart';
+import '../controllers/profile_controller.dart';
 
 class MainLayoutBinding extends Bindings {
   @override
@@ -15,5 +16,6 @@ class MainLayoutBinding extends Bindings {
     Get.lazyPut<CustomerController>(() => CustomerController());
     Get.lazyPut<ProductController>(() => ProductController());
     Get.lazyPut<DashboardController>(() => DashboardController());
+    Get.lazyPut<ProfileController>(() => ProfileController());
   }
 }