+++ lib/presentation/bindings/profil_controller.dart
@@ -1,0 +1,9 @@
+import 'package:get/get.dart';
+import 'package:tontine_app/presentation/controllers/profile_controller.dart';
+
+class ProfileBinding extends Bindings {
+  @override
+  void dependencies() {
+    Get.lazyPut<ProfileController>(() => ProfileController());
+  }
+}